refactor: modularize socketHandler into separate modules
- Split large 285-line socketHandler.js into 6 focused modules
- Created handlers for: user registration, messages, invitations, typing, video calls, disconnection
- Maintained all existing functionality and exports
- Improved code maintainability and testability
- Added comprehensive documentation
