+++ backend/functions/sockets/handlers/index.js
@@ -1,0 +1,11 @@
+/**
+ * Socket Handlers Index
+ * Exports all socket event handlers
+ */
+
+export { handleUserRegistration, getUserSocketMap } from './userRegistrationHandler.js';
+export { handleMessages } from './messageHandler.js';
+export { handleInvitations } from './invitationHandler.js';
+export { handleTypingIndicators } from './typingHandler.js';
+export { handleVideoCalls } from './videoCallHandler.js';
+export { handleDisconnection } from './disconnectHandler.js';
